+++ druid-shell/examples/shello.rs
@@ -14,12 +14,12 @@
 
 use std::any::Any;
 
-use druid_shell::kurbo::{Line, Rect, Vec2};
+use druid_shell::kurbo::{Line, Rect};
 use druid_shell::piet::{Color, RenderContext};
 
 use druid_shell::{
-    Application, Cursor, FileDialogOptions, FileSpec, HotKey, KeyEvent, KeyModifiers, Menu,
-    MouseEvent, SysMods, TimerToken, WinHandler, WindowBuilder, WindowHandle,
+    Application, Cursor, FileDialogOptions, FileSpec, HotKey, KeyEvent, Menu, MouseEvent, SysMods,
+    TimerToken, WinHandler, WindowBuilder, WindowHandle,
 };
 
 const BG_COLOR: Color = Color::rgb8(0x27, 0x28, 0x22);
@@ -71,8 +71,8 @@ impl WinHandler for HelloState {
         false
     }
 
-    fn wheel(&mut self, delta: Vec2, mods: KeyModifiers) {
-        println!("mouse_wheel {:?} {:?}", delta, mods);
+    fn wheel(&mut self, event: &MouseEvent) {
+        println!("mouse_wheel {:?}", event);
     }
 
     fn mouse_move(&mut self, event: &MouseEvent) {