+++ druid/src/window.rs
@@ -155,7 +155,7 @@ impl<T: Data> Window<T> {
         env: &Env,
     ) -> bool {
         match &event {
-            Event::MouseDown(e) | Event::MouseUp(e) | Event::MouseMove(e) => {
+            Event::MouseDown(e) | Event::MouseUp(e) | Event::MouseMove(e) | Event::Wheel(e) => {
                 self.last_mouse_pos = Some(e.pos)
             }
             Event::Internal(InternalEvent::MouseLeave) => self.last_mouse_pos = None,